% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate.R
\name{allocate}
\alias{allocate}
\title{Allocate energy flows to batteries}
\usage{
allocate(flow, pref_v, soc_v, cap_v)
}
\arguments{
\item{flow}{kWh to be put into the system}

\item{pref_v}{the preference to share flows}

\item{soc_v}{the current state of charge}

\item{cap_v}{the actual capacity of the battery}
}
\value{
Numeric vector, where the last number is the actual overflow
}
\description{
Given a number of factors, the function returns the
upadted flows to the batteries
}
