% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{backshift}
\alias{backshift}
\title{Shift energy consumption towards the past.}
\usage{
backshift(matrix, price, horizon, depreciation, cap, size)
}
\value{
A matrix, or an xts object if the index provided are a POSIXct object.
}
\description{
Shift energy consumption towards the past.
}
