% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{df_to_ts}
\alias{df_to_ts}
\alias{ts_to_df}
\title{Convert Data Frame to a Time Series object.}
\usage{
df_to_ts(df)

ts_to_df(ts)
}
\arguments{
\item{df}{Data frame where the first column is a POXIXct objet or convertible.}

\item{ts}{Time series}
}
\value{
Time series indexed by the first column of df.

Data frame where the fist column is a POSIXt object calle "datetime"
}
\description{
This is a random description.
}
\details{
These are random details. Will it work somehow?
I will know it soon.
}
