% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribute.R
\name{distribute}
\alias{distribute}
\title{Distribute energy flows to batteries.}
\usage{
distribute(flow, soc, vol, share = 1, level = 1, active = 1,
  eff = 1, cap = 0)
}
\arguments{
\item{flow}{Energy flow to be exchanged in the following time step. 
Positive for charge, negative for discharge.}

\item{soc}{Current state of charge.}

\item{vol}{Objective state of charge to be achieved.}

\item{share}{Double, indicating the priority to share the energy flow.}

\item{level}{Integer, expressing the priority to be resolved, starting with the highest.}

\item{active}{boolean, expressing whether the battery should be considered}

\item{eff}{Efiiciency of the energy flow.}

\item{cap}{Maximum energy that can be exchanged witht the battery}
}
\value{
An object containing the state of charge of each battery
}
\description{
Given a number of factors, the function returns the
}
\examples{
1+1
}
