% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OO_frame.R
\docType{class}
\name{e_frame}
\alias{e_frame}
\title{Define an e_frame object}
\format{An \code{\link{R6Class}} generator object}
\usage{
e_frame
}
\description{
Define an e_frame object
}
\section{Fields}{

\describe{
\item{\code{serveraddress}}{Stores address of your lightning server.}

\item{\code{sessionid}}{Stores id of your current session on the server.}

\item{\code{url}}{Stores url of the last visualization created by this object.}

\item{\code{autoopen}}{Checks if the server is automatically opening the visualizations.}

\item{\code{notebook}}{Checks if the server is in the jupyter notebook mode.}
}}

\section{Methods}{

\describe{
  \item{Documentation}{For full documentation of each method go to https://github.com/lightning-viz/lightining-r/}
  \item{\code{new(serveraddress)}}{This method is used to create object of this class with \code{serveraddress} as address of the server object is connecting to.}
  \item{\code{sethost(serveraddress)}}{This method changes server that you are contacting with to \code{serveraddress}.}
  \item{\code{createsession(sessionname = "")}}{This method creates new session on the server with optionally given name in \code{sessionname}.}
  \item{\code{usesession(sessionid)}}{This method changes currently used session on the server to the one with id given in \code{sessionid} parameter.}
}
}

\keyword{data}
