% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreshift.R
\name{foreshift}
\alias{foreshift}
\title{Shift demand to the future}
\usage{
foreshift(input_mtx, flex_step, input_vct, fit = ~1 * .demand)
}
\arguments{
\item{input_mtx}{list of flexible matrices}

\item{flex_step}{list of flexibility steps}

\item{input_vct}{list of input vectorized}

\item{fit}{formula to calulate the fit of the flexible demand}
}
\value{

}
\description{
Shift demand to the future
}
