% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_power.R
\name{kw_to_kwh}
\alias{kw_to_kwh}
\alias{kwh_to_kw}
\title{Switch between power and energy based objects}
\usage{
kw_to_kwh(df)

kwh_to_kw(df)
}
\arguments{
\item{df}{Data frame to be transformed}
}
\value{
Data frame where the columns have been time-transformed
to signify energy or power transfromations.
}
\description{
In energy analysis, data is sometimes provided
power-based, when it is needed energy based.
These functions allow to swith between the two modes.
}
