% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qhour}
\alias{qhour}
\title{Simple energy flows every 15 minutes}
\format{A data frame with 58001 rows and 4 variables, left aligned every 15 minutes:
\describe{
  \item{datetime}{time series column}
  \item{from_grid}{energy from grid, in kWh}
  \item{to_grid}{energy exported to the grid, in kWh}
  \item{solar_production}{total energy generated by the solar panels, in kWh}
  ...
}}
\usage{
qhour
}
\description{
A dataset as a time series that displays summarized data of
production and consumption of energy in a building.
}
\keyword{datasets}
