% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OO_storage.R
\docType{class}
\name{storage}
\alias{storage}
\title{Class to define a storage}
\format{An \code{\link{R6Class}} generator object}
\usage{
storage
}
\description{
Class to define a storage
}
\section{Methods}{

\describe{
  \item{\code{example_method(parameter_1 = 3)}}{This method uses \code{parameter_1} to ...}
}
}

\keyword{data}
