% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{allocate}
\alias{allocate}
\title{Allocate energy flows to batteries.}
\usage{
allocate(flow, soc, vol, share = as.numeric(c(1)), level = as.numeric(c(1)),
  active = as.logical(c(1)), eff = as.numeric(c(1)),
  cap = as.numeric(c(0)))
}
\arguments{
\item{flow}{Energy flow to be exchanged in the following time step .
Positive for charge, negative for discharge.}

\item{soc}{Current state of charge.}

\item{vol}{Objective state of charge to be achieved.}

\item{share}{Double, indicating the priority to share the energy flow}

\item{level}{Integer, expressing the priority to be resolved, starting with
the highest.}

\item{active}{boolean, expressing whether the battery should be considered}

\item{eff}{Efiiciency of the energy flow.}

\item{cap}{Maximum energy that can be exchanged witht the battery}
}
\value{
An object containing the state of charge of each battery, its final
energy flow, and a double containing the energy that was not shared
}
\description{
Given a number of factors, the function returns the
demand and updated state of charge of the batteries.
}
