% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OO.R
\docType{class}
\name{flex_mtx}
\alias{flex_mtx}
\title{Another example of an R6 class}
\format{An \code{\link{R6Class}} generator object}
\usage{
flex_mtx
}
\description{
Another example of an R6 class
}
\section{Methods}{

\describe{
  \item{\code{example_method(parameter_1 = 3)}}{This method uses \code{parameter_1} to ...}
}
}

\keyword{data}
