% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{foreshift}
\alias{foreshift}
\title{Shift energy consumption towards the future.}
\usage{
foreshift(matrix, flex_step, cap = 0, cap_spread = TRUE, foresee = TRUE,
  solar = as.numeric(c(0)))
}
\arguments{
\item{matrix}{Numeric matrix, each row an observation in time (it can be a time series)
and each column representing the flexibility of demand, increasing towards the right.}

\item{flex_step}{Integer vector of same length of the column of `matrix`, indicating
the number of timesteps the demand can be delayed. Normally, the first integer is `0`,
indicating the `solid` demand.}

\item{cap}{If higher than 0, indicates the maximum capacity of flexible demand
that can be allocated in a timestep.}

\item{cap_spread}{Boolean. if true, the cap is never exceeded, and instead the
consumption is displaced towards the future, and resolved as soon as possible.}

\item{foresee}{Boolean. If false, it indicates a complete lack of knowledge of the
future flexible demand in the timeline.}

\item{solar}{If present, the fitting curve is calculated substracting solar
from the demand.}
}
\value{
A matrix, or an xts object if the index provided are a POSIXct object.
}
\description{
In function of a fitting curve and the flexibility time, the consumption of
energy is delayed.
}
