% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreshift.R
\name{foreshift}
\alias{foreshift}
\title{Shift demand to the future}
\usage{
foreshift(input_mtx, flex_step, cap = NULL, input_vct, fit = ~1 *
  .demand)
}
\arguments{
\item{input_mtx}{list of flexible matrices}

\item{flex_step}{list of flexibility steps}

\item{cap}{list of cap per object}

\item{input_vct}{list of input vectorized}

\item{fit}{formula to calulate the fit of the flexible demand}
}
\value{
object 
\describe{
  \item{demand fixed}{First item}
  \item{demand flex}{Second item}
  \item{fit curve initial}{Second item}
  \item{fit curve final}{Second item}   
}
}
\description{
\tabular{rrrrr}{
  21.0 \tab 6 \tab 160 \tab 110 \tab 3.90\cr
  21.0 \tab 6 \tab 160 \tab 110 \tab 3.90\cr
  22.8 \tab 4 \tab 108 \tab  93 \tab 3.85\cr
  21.4 \tab 6 \tab 258 \tab 110 \tab 3.08\cr
  18.7 \tab 8 \tab 360 \tab 175 \tab 3.15
}
}
\examples{
1+1
}
